chore: consolidate CopyToClipboard and clipboard utilities to @superset-ui/core
This change:
- Creates a generic CopyToClipboard component in @superset-ui/core
  - Uses onSuccess/onError callbacks (not tied to any toast system)
  - Includes tooltip, text display, and async getText support
- Creates clipboard utilities in @superset-ui/core
  - copyTextToClipboard with Safari/Chrome compatibility
  - isSafari browser detection utility
- Creates ErrorBoundary in @superset-ui/core (removes react-error-boundary dep)
- Updates src/components/CopyToClipboard to be a thin wrapper that:
  - Integrates with Superset's toast notification system
  - Delegates to the core component for actual functionality
- Updates src/utils/copy.ts to re-export from core for backward compatibility

Developers can now:
- Use the generic CopyToClipboard from @superset-ui/core/components directly
- Use the toast-integrated version from src/components for Superset-specific use
