+++ superset-frontend/packages/superset-ui-core/src/chart/index.ts
@@ -26,6 +26,11 @@ export { ChartProps };
 export type { ChartPropsConfig };
 
 export { default as createLoadableRenderer } from './components/createLoadableRenderer';
+export {
+  ErrorBoundary,
+  type ErrorBoundaryProps,
+  type FallbackProps,
+} from './components/ErrorBoundary';
 export { default as reactify } from './components/reactify';
 export { default as SuperChart } from './components/SuperChart';
